style: Alert Context, Dialog, Provider 코드 분리
